+++ server/Models/Domain/InventoryTO.cs
@@ -10,7 +10,7 @@ namespace server.Models.Domain
             Quantity = quantity;
             DateLastUpdated = dateLastUpdated;
         }
-        public int SKU { get; set; }
+        public double SKU { get; set; }
         public int InventoryID { get; set; }
         public int Quantity { get; set; }
         public DateTime DateLastUpdated { get; set; }
