+++ server/Services/IInventoryService.cs
@@ -7,10 +7,10 @@ namespace server.Services
     public interface IInventoryService
     {
         Task CreateInventory(InventoryTO inventory);
-        Task<int> GetQuantity(int SKU);
+        Task<double> GetQuantity(double SKU);
         Task<DateTime> GetDateLastUpdated(int inventoryId);
-        Task SetQuantity(int SKU, int amount);
-        Task AddQuantity(int SKU, int amount);
-        Task SubtractQuantity(int SKU, int amount);
+        Task SetQuantity(double SKU, int amount);
+        Task AddQuantity(double SKU, int amount);
+        Task SubtractQuantity(double SKU, int amount);
     }
 }
